Define headers for distribution
