+++ src/Makefile.am
@@ -1,11 +1,9 @@
 lib_LTLIBRARIES = libdavenport.la
-include_HEADERS = \
-  aggregate_solution.h \
-  davenport.h \
-  lower_bound.h \
-  network.h \
-  preference_graph.h \
-  ranking.h
+nobase_include_HEADERS = \
+  davenport/aggregate_solution.h \
+  davenport/davenport.h \
+  davenport/network.h \
+  davenport/preference_graph.h
 
 libdavenport_la_SOURCES = \
   aggregate_solution.c aggregate_solution.h \
